requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
matplotlib>=3.4.0
tqdm>=4.60.0
Pillow>=8.2.0
lpips>=0.1.4
pytorch-fid>=0.2.1
tensorboard>=2.6.0
PyYAML>=6.0
scikit-image>=0.18.0
scipy>=1.7.0
einops>=0.3.2